# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Maps Tracker API
  description: Real-time GPS vehicle tracking application with location history and place of interest management
  version: 1.0.0
  contact:
    name: Maps Tracker Support
    url: https://github.com/your-org/maps-tracker
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        username:
          type: string
          description: Unique username
        email:
          type: string
          format: email
          description: Email address
        role:
          type: string
          enum: [admin, manager, viewer]
          description: User role
        is_active:
          type: boolean
          description: Whether user account is active
        must_change_password:
          type: boolean
          description: Whether user must change password on next login
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
      required:
        - id
        - username
        - email
        - role
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
      required:
        - error
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource already exists"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Too many requests"
            remaining_attempts: 2
security:
  - cookieAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  description: Password with at least 8 characters, 1 uppercase, 1 number, 1 symbol
              required:
                - username
                - email
                - password
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
        '400':
          description: Invalid input or user already exists
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and create session
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Too many login attempts
          $ref: '#/components/responses/TooManyRequests'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: End current user session
      operationId: logoutUser
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password
      description: Update user password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
              required:
                - current_password
                - new_password
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...